paper/requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-login==0.6.2
flask-principal==0.4.0
flask-bcrypt==1.0.1
flask-wtf==1.1.1
redis==4.5.4
pymysql==1.0.3
python-dotenv==1.0.0
matplotlib==3.7.1
pdfkit==1.0.0
email_validator==2.0.0
Pillow==9.5.0 